requirements.txt
h11
pillow
pip
distro
wheel
setuptools
colorama
sniffio
markupsafe
werkzeug
click
jinja2
blinker
itsdangerous
httpcore
idna
httpx
Brotli
jiter
openai
pydantic
anyio
zopfli
certifi
webencodings
fonttools
pycparser
tinycss2
packaging
tomli
autocommand
inflect
typeguard
platformdirs
zipp
cssselect2
pydyf
cffi
pyphen
tinyhtml5
flask
python-dotenv